--- a/.pre-commit-config.yaml
+++ b/.pre-commit-config.yaml
@@ -54,7 +54,7 @@ repos:
     rev: v3.0.0
     hooks:
       - id: pyupgrade
-        args: [ "--py36-plus", "--py37-plus", "--py38-plus", "--py39-plus", "--py310-plus" ]
+        args: [ "--py36-plus", "--py37-plus" ]
 
   - repo: https://github.com/pre-commit/mirrors-mypy
     rev: v0.982