--- a/src/token_bucket.rs
+++ b/src/token_bucket.rs
@@ -41,8 +41,7 @@ impl ThreadState {
 /// Async context manager useful for controlling client traffic
 /// in situations where you need to limit traffic to `n` requests per `m` unit of time.
 /// For example, when you can only send 1 request per minute.
-#[pyclass]
-// #[pyclass(frozen)]  # <-- TODO: Add in when on pyo3 v0.17
+#[pyclass(frozen)]
 #[pyo3(name = "TokenBucket")]
 #[pyo3(module = "self_limiters")]
 pub struct TokenBucket {