Upgrades to pyo3 v17 and implements the frozen attribute that was included in the new release. Will see if we can benchmark before and after to see the difference.