@sondrelg 2022-10-12
Benchmark before:

```
Testing token bucket...
Discarding 3 first samples
Average was 0.3ms per run
Median was 0.3ms per run
Median time was below target of 0.4ms

Testing semaphore...
Discarding 3 first samples
Average was 3.2ms per run
Median was 3.4ms per run
Median time was below target of 4.0ms
```

Benchmark after (worse)

```
Testing token bucket...
Discarding 3 first samples
Average was 0.3ms per run
Median was 0.3ms per run
Median time was below target of 0.4ms

Testing semaphore...
Discarding 3 first samples
Average was 3.5ms per run
Median was 3.9ms per run
Median time was below target of 4.0ms
```

Benchmark after, using release build

```
Testing token bucket...
Discarding 3 first samples
Average was 0.2ms per run
Median was 0.2ms per run
Median time was below target of 0.4ms

Testing semaphore...
Discarding 3 first samples
Average was 2.4ms per run
Median was 2.6ms per run
Median time was below target of 4.0ms
```